requirements.txt
# Core Dependencies
streamlit>=1.31.0
google-generativeai>=0.3.0
pandas>=2.0.0
openpyxl>=3.1.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Additional utilities
Pillow>=10.0.0
